Wrong type

+++ src/sensors/bno080sensor.h
@@ -53,7 +53,7 @@ private:
     uint8_t lastReset = 0;
     BNO080Error lastError{};
 
-    uint8_t lastCalib = 0;
+    unsigned long lastCalib = 0;
     bool calibStopped = false;
 
     // Magnetometer specific members
